require multi-methods to use default methods
by handling dispatch for multimethods qlkit was not calling
multimethods that didn’t have a function for the current query. This
firstly required that mutimethods be used and resulted in behavior
where the multi method wouldn’t fail when it’s default method was
missing.

Better to use multi-methods as they are and allow the use of a simple
function for dispatch.  This behavior is simpler and easier to intuit.
While it may be less convenient as you have to create a default
dispatch function to get with missing key handlers.